Moved the Fail API into its own package.
