+++ Core/src/main/java/de/uka/ipd/sdq/beagle/core/failurehandling/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Beagle’s handling of exceptional situations.
+ */
+package de.uka.ipd.sdq.beagle.core.failurehandling;
